Module_1/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
segment-anything>=1.0
ultralytics>=8.0.0
scikit-image>=0.21.0
matplotlib>=3.7.0
scipy>=1.11.0
transformers>=4.30.0
